Create movable walls
Create a new class MovableWall, which extends wall. After MovableWall is instantiated with init(), set_movement_properties() should be called on it to set the settings of its movement. if the block does not have the moving_immediately parameter in movement_properties() set to true, it will need to be activated by calling trigger_movement(true) to start moving