Door determines next level property
User would need to click the interact button 'I' when the player is in the door in order to load the next level. Test level is level 8.

Please review how I am doing the flag for interaction as I am not sure if there's a better way to check for it